Fix Migration
- Исправлены баги которые вызывали критические ошибки связанные с миграцией
- Небольшой рефакторинг кода
- Метод фабрики панелей отвечающий за панели событий теперь возвращает объект EventPanel

+++ Assets/1 - Transmigratio/Scripts/Gameplay/Scenarios/MathModel/Demography.cs
@@ -139,7 +139,7 @@ namespace Gameplay.Scenarios {
             public float prodModeK_S = 0.65f;
             public float governmentCorruption_L = 0.4f;
             public float governmentCorruption_M = 0.45f;
-            public float MinPiecePopulation = 50;
+            public int MinPiecePopulation = 50;
             public float val1 = 50f;
             public float val2 = 60f;
             public float val3 = 50f;
